docker/writer-app/app/requirements.txt
redis
cassandra-driver
